271 - Setup proper component structure
Applied general pattern of cleanup for consistency
across the application. A lot of files and folders
have changed and structure alike. Cleanup effort spans
across all the files in the frontend/src directory.
