+++ frontend/src/common/ColorPicker.jsx
@@ -1,7 +1,15 @@
+import { Typography } from '@mui/material';
 import { MuiColorInput } from 'mui-color-input';
 
-const ColorPicker = ({ value, handleChange }) => {
-  return <MuiColorInput fullWidth format="hex" value={value} onChange={handleChange} label="Assign Color" />;
+const ColorPicker = ({ label, value, handleChange }) => {
+  return (
+    <>
+      <Typography variant="subtitle2" color="text.secondary">
+        {label}
+      </Typography>
+      <MuiColorInput size="small" fullWidth format="hex" value={value} onChange={handleChange} />
+    </>
+  );
 };
 
 export default ColorPicker;
