+++ frontend/src/features/Categories/categoriesSaga.js
@@ -1,7 +1,8 @@
-import { takeLatest, put, call } from 'redux-saga/effects';
+import { call, put, takeLatest } from 'redux-saga/effects';
+
+import instance from '../../utils/Instances';
 import { categoryActions } from './categoriesSlice';
 import { REACT_APP_LOCALHOST_URL } from '../../utils/Common';
-import instance from '../../utils/Instances';
 
 const DEFAULT_LIMIT = 10;
 const BASEURL = `${REACT_APP_LOCALHOST_URL}/api/v1`;
@@ -20,35 +21,6 @@ export function* getCategories(action) {
     yield put(categoryActions.getCategoriesSuccess(response.data));
   } catch (e) {
     yield put(categoryActions.getCategoriesFailure(e));
-  }
-}
-
-export function* getCategory(action) {
-  try {
-    const catID = action.payload;
-    const userID = localStorage.getItem('userID');
-    const params = new URLSearchParams();
-    params.append('id', userID);
-    params.append('catID', catID);
-    const response = yield call(instance.get, `${BASEURL}/category?${params.toString()}`);
-    yield put(categoryActions.getCategorySuccess(response.data));
-  } catch (e) {
-    yield put(categoryActions.getCategoryFailure(e));
-  }
-}
-
-export function* getItemsForCategory(action) {
-  try {
-    const catID = action.payload;
-    const userID = localStorage.getItem('userID');
-    const params = new URLSearchParams();
-    params.append('id', userID);
-    params.append('catID', catID);
-    params.append('limit', DEFAULT_LIMIT);
-    const response = yield call(instance.get, `${BASEURL}/category/items?${params.toString()}`);
-    yield put(categoryActions.getItemsForCategorySuccess(response.data));
-  } catch (e) {
-    yield put(categoryActions.getItemsForCategoryFailure(e));
   }
 }
 
@@ -81,49 +53,6 @@ export function* removeCategory(action) {
   }
 }
 
-export function* fetchAddItemsInCategory(action) {
-  try {
-    const userID = localStorage.getItem('userID');
-    const { id, rowSelected, collaborators } = action.payload;
-    const response = yield call(instance.post, `${BASEURL}/category/items`, {
-      id,
-      userID,
-      assetIDs: rowSelected,
-      collaborators: collaborators,
-    });
-    yield put(categoryActions.addItemsInCategorySuccess(response.data));
-  } catch (e) {
-    yield put(categoryActions.addItemsInCategoryFailure(e));
-  }
-}
-
-export function* uploadImage(action) {
-  try {
-    const { id, selectedImage } = action.payload;
-    const formData = new FormData();
-    formData.append('imageSrc', selectedImage);
-    const response = yield call(instance.post, `${BASEURL}/${id}/uploadImage`, formData);
-    yield put(categoryActions.uploadImageSuccess(response.data));
-  } catch (e) {
-    yield put(categoryActions.uploadImageFailure(e));
-  }
-}
-
-export function* getSelectedImage(action) {
-  try {
-    const { id } = action.payload;
-    // we need to modify the image to be of arrayBuffer type and build a blob object from it
-    const response = yield call(instance.get, `${BASEURL}/${id}/fetchImage`, {
-      responseType: 'arraybuffer',
-    });
-    const blob = new Blob([response.data], { type: response.headers['content-type'] });
-    const avatarUrl = URL.createObjectURL(blob);
-    yield put(categoryActions.getSelectedImageSuccess(avatarUrl));
-  } catch (e) {
-    yield put(categoryActions.getSelectedImageFailure(e));
-  }
-}
-
 export function* download() {
   try {
     const userID = localStorage.getItem('userID');
@@ -140,14 +69,6 @@ export function* watchGetCategoryList() {
   yield takeLatest(`category/getCategories`, getCategories);
 }
 
-export function* watchGetCategory() {
-  yield takeLatest(`category/getCategory`, getCategory);
-}
-
-export function* watchGetItemsForCategory() {
-  yield takeLatest(`category/getItemsForCategory`, getItemsForCategory);
-}
-
 export function* watchCreateCategory() {
   yield takeLatest(`category/createCategory`, createCategory);
 }
@@ -160,31 +81,8 @@ export function* watchRemoveCategory() {
   yield takeLatest(`category/removeCategory`, removeCategory);
 }
 
-export function* watchFetchAddItemsInCategory() {
-  yield takeLatest(`category/addItemsInCategory`, fetchAddItemsInCategory);
-}
-
-export function* watchUploadImage() {
-  yield takeLatest(`category/uploadImage`, uploadImage);
-}
-
-export function* watchGetSelectedImage() {
-  yield takeLatest(`category/getSelectedImage`, getSelectedImage);
-}
-
 export function* watchDownload() {
   yield takeLatest(`category/download`, download);
 }
 
-export default [
-  watchGetCategoryList,
-  watchGetCategory,
-  watchDownload,
-  watchGetItemsForCategory,
-  watchCreateCategory,
-  watchUpdateCategory,
-  watchRemoveCategory,
-  watchFetchAddItemsInCategory,
-  watchUploadImage,
-  watchGetSelectedImage,
-];
+export default [watchGetCategoryList, watchDownload, watchCreateCategory, watchUpdateCategory, watchRemoveCategory];
