+++ frontend/src/features/LandingPage/HeroContent/FloatingBarChart.jsx
