+++ frontend/src/features/MaintenancePlanDetails/MaintenancePlanItem/ItemContent/ItemGraph.jsx
@@ -1,0 +1,27 @@
+import { Box } from '@mui/material';
+import BarChart from '../../../../common/Chart/BarChart';
+import RowHeader from '../../../../common/RowHeader';
+
+export default function ItemGraph({ totalItems }) {
+  return (
+    <>
+      <RowHeader title="Graph" caption="Graph details for last 10 recently updated" />
+      <Box sx={{ position: 'relative', width: 'calc(100% - 1rem)' }}>
+        <BarChart
+          legendLabel="Name Vs Cost"
+          data={
+            totalItems
+              ?.filter((_, index) => index < 10)
+              ?.map((v, index) => ({
+                label: v.name,
+                count: v.price,
+                color: ['rgb(54, 162, 235)', 'rgb(211, 211, 211)'][index % 2],
+              })) || []
+          }
+          backgroundColor="rgba(75, 192, 192, 0.4)"
+          borderColor="rgba(75, 192, 192, 1)"
+        />
+      </Box>
+    </>
+  );
+}
