+++ frontend/src/features/MaintenancePlanList/maintenanceSaga.js
@@ -1,7 +1,8 @@
-import { takeLatest, put, call } from 'redux-saga/effects';
-import { REACT_APP_LOCALHOST_URL } from '../../utils/Common';
+import { call, put, takeLatest } from 'redux-saga/effects';
+
 import instance from '../../utils/Instances';
 import { maintenancePlanActions } from './maintenanceSlice';
+import { REACT_APP_LOCALHOST_URL } from '../../utils/Common';
 
 const DEFAULT_LIMIT = 10;
 const MAINTENANCE_STATUS_OPTION_TYPE = 'maintenance';
@@ -68,50 +69,6 @@ export function* removePlan(action) {
   }
 }
 
-export function* getSelectedMaintenancePlan(action) {
-  try {
-    const mID = action.payload;
-    const userID = localStorage.getItem('userID');
-    const params = new URLSearchParams();
-    params.append('id', userID);
-    params.append('mID', mID);
-    const response = yield call(instance.get, `${BASEURL}/plan?${params.toString()}`);
-    yield put(maintenancePlanActions.getSelectedMaintenancePlanSuccess(response.data));
-  } catch (e) {
-    yield put(maintenancePlanActions.getSelectedMaintenancePlanFailure(e));
-  }
-}
-
-export function* getItemsInMaintenancePlan(action) {
-  try {
-    const mID = action.payload;
-    const userID = localStorage.getItem('userID');
-    const params = new URLSearchParams();
-    params.append('id', userID);
-    params.append('mID', mID);
-    params.append('limit', DEFAULT_LIMIT);
-    const response = yield call(instance.get, `${BASEURL}/plans/items?${params.toString()}`);
-    yield put(maintenancePlanActions.getItemsInMaintenancePlanSuccess(response.data));
-  } catch (e) {
-    yield put(maintenancePlanActions.getItemsInMaintenancePlanFailure(e));
-  }
-}
-
-export function* fetchAddItemsInPlan(action) {
-  try {
-    const userID = localStorage.getItem('userID');
-    const { id, rowSelected, collaborators } = action.payload;
-    const response = yield call(instance.post, `${BASEURL}/plans/items`, {
-      id,
-      userID,
-      assetIDs: rowSelected,
-      collaborators: collaborators,
-    });
-    yield put(maintenancePlanActions.addItemsInPlanSuccess(response.data));
-  } catch (e) {
-    yield put(maintenancePlanActions.addItemsInPlanFailure(e));
-  }
-}
 
 export function* download() {
   try {
@@ -122,33 +79,6 @@ export function* download() {
     yield put(maintenancePlanActions.downloadSuccess(response.data));
   } catch (e) {
     yield put(maintenancePlanActions.downloadFailure(e));
-  }
-}
-
-export function* uploadImage(action) {
-  try {
-    const { id, selectedImage } = action.payload;
-    const formData = new FormData();
-    formData.append('imageSrc', selectedImage);
-    const response = yield call(instance.post, `${BASEURL}/${id}/uploadImage`, formData);
-    yield put(maintenancePlanActions.uploadImageSuccess(response.data));
-  } catch (e) {
-    yield put(maintenancePlanActions.uploadImageFailure(e));
-  }
-}
-
-export function* getSelectedImage(action) {
-  try {
-    const { id } = action.payload;
-    // we need to modify the image to be of arrayBuffer type and build a blob object from it
-    const response = yield call(instance.get, `${BASEURL}/${id}/fetchImage`, {
-      responseType: 'arraybuffer',
-    });
-    const blob = new Blob([response.data], { type: response.headers['content-type'] });
-    const avatarUrl = URL.createObjectURL(blob);
-    yield put(maintenancePlanActions.getSelectedImageSuccess(avatarUrl));
-  } catch (e) {
-    yield put(maintenancePlanActions.getSelectedImageFailure(e));
   }
 }
 
@@ -172,29 +102,10 @@ export function* watchRemovePlan() {
   yield takeLatest(`maintenancePlan/removePlan`, removePlan);
 }
 
-export function* watchGetSelectedMaintenancePlan() {
-  yield takeLatest(`maintenancePlan/getSelectedMaintenancePlan`, getSelectedMaintenancePlan);
-}
-
-export function* watchGetItemsInMaintenancePlan() {
-  yield takeLatest(`maintenancePlan/getItemsInMaintenancePlan`, getItemsInMaintenancePlan);
-}
-
-export function* watchFetchAddItemsInPlan() {
-  yield takeLatest(`maintenancePlan/addItemsInPlan`, fetchAddItemsInPlan);
-}
-
 export function* watchDownload() {
   yield takeLatest(`maintenancePlan/download`, download);
 }
 
-export function* watchUploadImage() {
-  yield takeLatest(`maintenancePlan/uploadImage`, uploadImage);
-}
-
-export function* watchGetSelectedImage() {
-  yield takeLatest(`maintenancePlan/getSelectedImage`, getSelectedImage);
-}
 
 export default [
   watchGetPlanList,
@@ -203,9 +114,4 @@ export default [
   watchDownload,
   watchUpdatePlan,
   watchRemovePlan,
-  watchFetchAddItemsInPlan,
-  watchGetItemsInMaintenancePlan,
-  watchGetSelectedMaintenancePlan,
-  watchGetSelectedImage,
-  watchUploadImage,
 ];
