+++ frontend/src/features/Notes/AddNote/AddNote.jsx
@@ -1,17 +1,16 @@
-import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
-import { DatePicker } from '@mui/x-date-pickers/DatePicker';
 import { enqueueSnackbar } from 'notistack';
 import { useEffect, useState } from 'react';
-import { Button, FormControl, InputLabel, MenuItem, Select, Stack, TextField } from '@mui/material';
-import { ADD_NOTES_FORM_FIELDS, STATUS_OPTIONS } from './constants';
+import { Button, Stack } from '@mui/material';
+import { ADD_NOTES_FORM_FIELDS, STATUS_OPTIONS } from '../constants';
 import { useDispatch } from 'react-redux';
 import { AddRounded, CheckCircleRounded } from '@mui/icons-material';
-import { notesActions } from './notesSlice';
-import ColorPicker from '../../common/ColorPicker';
-import RetrieveUserLocation from '../../common/Location/RetrieveUserLocation';
-import LocationPicker from '../../common/Location/LocationPicker';
-import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider/LocalizationProvider';
+import { notesActions } from '../notesSlice';
+import ColorPicker from '../../../common/ColorPicker';
 import dayjs from 'dayjs';
+import AddNoteHeader from './AddNoteHeader';
+import AddNoteStatusOptions from './AddNoteStatusOptions';
+import CustomDatePicker from '../../../common/DatePicker/CustomDatePicker';
+import AddNoteLocationPicker from './AddNoteLocationPicker';
 
 const AddNote = ({ setEditMode, setSelectedNoteID, noteID, notes }) => {
   const dispatch = useDispatch();
@@ -125,92 +124,23 @@ const AddNote = ({ setEditMode, setSelectedNoteID, noteID, notes }) => {
   }, [noteID]);
 
   return (
-    <Stack spacing="1rem">
-      <Stack direction="row" spacing="1rem" alignItems="center" flexGrow="1">
-        {Object.values(formFields)
-          .slice(0, 1)
-          .map((v, index) => (
-            <TextField
-              key={index}
-              id={v.name}
-              name={v.name}
-              label={v.label}
-              value={v.value}
-              placeholder={v.placeholder}
-              onChange={handleInput}
-              required={v.required}
-              fullWidth={v.fullWidth}
-              error={!!v.errorMsg}
-              helperText={v.errorMsg}
-              variant={v.variant}
-              minRows={v.rows || 4}
-              multiline={v.multiline || false}
-            />
-          ))}
-        <RetrieveUserLocation setLocation={setLocation} />
-      </Stack>
-      <Stack
-        sx={{
-          flexDirection: { xs: 'column', sm: 'row' },
-          alignItems: 'center',
-          gap: '1rem',
-        }}
-      >
-        {Object.values(formFields)
-          .slice(1)
-          .map((v, index) => (
-            <TextField
-              key={index}
-              id={v.name}
-              name={v.name}
-              label={v.label}
-              value={v.value}
-              placeholder={v.placeholder}
-              onChange={handleInput}
-              required={v.required}
-              fullWidth={v.fullWidth}
-              error={!!v.errorMsg}
-              helperText={v.errorMsg}
-              variant={v.variant}
-              minRows={v.rows || 4}
-              multiline={v.multiline || false}
-            />
-          ))}
-      </Stack>
-      <FormControl fullWidth>
-        <InputLabel id="status-selector-label">Selected status</InputLabel>
-        <Select
-          labelId="status-selector-labelId"
-          id="status-selector"
-          value={status}
-          name={'status'}
-          onChange={handleStatus}
-          variant="standard"
-        >
-          {STATUS_OPTIONS.map((option) => (
-            <MenuItem key={option.id} value={option.label}>
-              {option.display}
-            </MenuItem>
-          ))}
-        </Select>
-      </FormControl>
-      <ColorPicker value={planColor} handleChange={handleColorChange} />
-      <LocalizationProvider dateAdapter={AdapterDayjs}>
-        <DatePicker
-          label={'Estimated completion date'}
-          disablePast
-          value={completionDate}
-          onChange={setCompletionDate}
-        />
-      </LocalizationProvider>
-      {location.lat ? (
-        <LocationPicker subtitle="Select location" location={location} onLocationChange={setLocation} editMode={true} />
-      ) : null}
-      <Button
-        onClick={submit}
-        startIcon={noteID ? <CheckCircleRounded /> : <AddRounded />}
-        sx={{ alignSelf: 'flex-start' }}
-      >
+    <Stack spacing={1}>
+      <AddNoteHeader formFields={formFields} handleInput={handleInput} setLocation={setLocation} />
+      <AddNoteStatusOptions label="Selected status" name="status" value={status} handleStatus={handleStatus} />
+      <ColorPicker label="Assign Color" value={planColor} handleChange={handleColorChange} />
+      <CustomDatePicker
+        label="Assign estimated completion date"
+        completionDate={completionDate}
+        setCompletionDate={setCompletionDate}
+      />
+      <AddNoteLocationPicker
+        subtitle="Select location"
+        location={location}
+        setLocation={setLocation}
+        editMode={true}
+        displayLocationPicker={location?.lat}
+      />
+      <Button onClick={submit} variant="outlined" startIcon={noteID ? <CheckCircleRounded /> : <AddRounded />}>
         {noteID ? 'Save' : 'Add'}
       </Button>
     </Stack>
