+++ frontend/src/features/Notes/constants.jsx
@@ -27,49 +27,49 @@ export const STATUS_OPTIONS = [
   {
     id: 1,
     label: 'draft',
-    display: 'Draft',
+    display: 'Draft Status',
     icon: <DraftsRounded fontSize="small" color="primary" />,
   },
   {
     id: 2,
     label: 'archived',
-    display: 'Archived',
+    display: 'Archived Status',
     icon: <ArchiveRounded fontSize="small" color="secondary" />,
   },
   {
     id: 3,
     label: 'completed',
-    display: 'Completed',
+    display: 'Completed Status',
     icon: <CheckCircleRounded fontSize="small" color="primary" />,
   },
   {
     id: 4,
     label: 'pending',
-    display: 'Pending',
+    display: 'Pending Status',
     icon: <PendingActionsRounded fontSize="small" color="secondary" />,
   },
   {
     id: 5,
     label: 'urgent',
-    display: 'Urgent',
+    display: 'Urgent Status',
     icon: <AnnouncementRounded fontSize="small" color="info" />,
   },
   {
     id: 6,
     label: 'general',
-    display: 'General',
+    display: 'General Status',
     icon: <NoteRounded fontSize="small" color="info" />,
   },
   {
     id: 7,
     label: 'on_hold',
-    display: 'On Hold',
+    display: 'On Hold Status',
     icon: <PauseRounded fontSize="small" color="warning" />,
   },
   {
     id: 8,
     label: 'cancelled',
-    display: 'Cancelled',
+    display: 'Cancelled Status',
     icon: <CancelRounded fontSize="small" color="error" />,
   },
 ];
@@ -80,6 +80,7 @@ export const ADD_NOTES_FORM_FIELDS = {
     placeholder: 'Short note title',
     value: '',
     name: 'title',
+    size: 'small',
     errorMsg: '',
     required: true,
     fullWidth: true,
@@ -100,6 +101,7 @@ export const ADD_NOTES_FORM_FIELDS = {
     placeholder: 'Note description in less than 500 characters',
     value: '',
     name: 'description',
+    size: 'small',
     errorMsg: '',
     required: false,
     fullWidth: true,
