+++ frontend/vite.config.js
@@ -14,7 +14,8 @@ export default defineConfig({
   },
   server: {
     watch: {
-     usePolling: true,
+      usePolling: true,
     },
     host: true,
-  }});
+  },
+});
